utxo_grabber/requirements.txt
kafka-python==2.0.2
redis==4.2.2
requests==2.27.1
splitstream==1.2.4
watchdog==2.1.7
